move `GroupingContext` to separate module
