+++ tests/sentry/grouping/test_strategies.py
@@ -3,7 +3,8 @@ from typing import Any
 import pytest
 
 from sentry.grouping.component import StacktraceGroupingComponent
-from sentry.grouping.strategies.base import GroupingContext, create_strategy_configuration_class
+from sentry.grouping.context import GroupingContext
+from sentry.grouping.strategies.base import create_strategy_configuration_class
 from sentry.services.eventstore.models import Event
 from sentry.testutils.cases import TestCase
 from sentry.testutils.helpers.eventprocessing import save_new_event
